Integrate translation files into final build
A longstanding bug has centered around the location of the translation files.
If Able Player is unable to find the files (which reside by default in the
/translations directory), this generates a fatal error.

The data-translation-path attribute had been added as a workaround.

The current fix integrates the translation files into the getTextTranslation()
function, so Able Player no longer needs to find, open, and read the
external files.