Regarding Issue #205 https://github.com/ableplayer/ableplayer/issues/205
I used "local_es.html" for my testing.

I added a second chapter track file called "wwa_chapters_es.vtt" and
included the source in video element in "local_es.html". Also, both
chapter tracks now have a specified "scrlang". The only difference between
the English chapters and the Spanish ones is that "Intro" becomes
"Introducción" in Spanish. This is not to say that "Intro" would be an
invalid abbreviation in Spanish, I just needed to make some distinction.

The chapter tracks are handled as follows:

1. If the captions are on, the chapter track tries to match the language
of the captions.
2. If the captions are off, but the transcript is on, the chapter track
tries to match the transcript.
3. If both the captions and the transcript are off (not showing), then the
chapter track tries to match the player.
4. If no matches are made in any of the above situations, it defaults to
the first chapter track available.

The code implementing that logic can be found in the
"updateChaptersLanguage" function. (see chapters.js)

The rest of the code:

1. I created  "updateChaptersLanguage" to implement the logic as I
described above. (see chapters.js)

2. In "setupPopups" I added "updateChaptersLanguage" under the condition '
else if (popup == 'chapters') '.
That way it always tries to match the appropriate language before changing
the text in the chapter pop-ups. (see buildplayer.js)

3. Because that is all is inside "setupPopups" it will be called by
"recreatePlayer" when the player initially builds. (see
ableplayer-base.js)

4. I added "setupPopups('chapters')" into "refreshControls". This makes
the chapter track respond to the captions because "refreshControls()" is
called in both "getCaptionClickFunction" and "getCaptionOffFunction". (see
control.js)

5. To respond to changes in the language of the transcript, I added
"thisObj.setupPopups ('chapters');" to the callback in
"$transcriptLanguageSelect.change()". (see transcript.js)

6. To respond to the trascripts being toggled on and off, I added
"this.setupPopups('chapters');" to the bottom of "handleTranscriptToggle".
(see control.js)