Try again to add lang to Supported Languages in README
GitHub-flavored Markdown doesn't have anything comparable to the
HTML lang attribute, so the list of languages needs to be coded in HTML.
My previous commit to address this issue include SPAN tags with
lang attributes, but when this is rendered as HTML the SPAN tags are
stripped out. Trying again with STRONG tags.